+++ mysql_perf.py
@@ -9,43 +9,52 @@
         formats or to the database.
 
     Usage:
-        mysql_perf.py -c file -d path {-S} [-j [-f]] [-n count] [-b seconds]
-            [-i [db_name:table_name] -m file]
-            [-o [dir_path]/file [-a]] [-y flavor_id] [-z]
+        mysql_perf.py -c file -d path
+            {-S [-j [-f]] [-n count] [-b seconds]
+                [-t email_addr [email_addr2 ...] [-s subject_line] [-u]]
+                [-i [db_name:table_name] -m file]
+                [-o [dir_path]/file [-a]] [-w] [-z]}
+            [-y flavor_id]
             [-v | -h]
 
     Arguments:
         -c file => MySQL server configuration file.  Required arg.
         -d dir path => Directory path to config file (-c). Required arg.
+
         -S => MySQL Database Performance Statistics option.
-        -j => Return output in JSON format.
-        -f => Flatten the JSON data structure to file and standard out.
-            For use with the -j option.
-        -n {count} => Number of loops to run the program.
-            Default:  1
-        -b {seconds} => Polling interval in seconds.
-            Default:  1
-        -i {database:collection} => Name of database and collection to insert
-            the database statistics data into.
-            Requires -m options.
-            Default:  sysmon:mysql_perf
-        -m file => Mongo config file.  Is loaded as a python, do not include
-            the .py extension with the name.
-            Required for -i option.
-        -o [path]/file => Directory path and file name for output.
-        -a => Append output to output file.
+            -j => Return output in JSON format.
+                -f => Flatten the JSON data structure to file and standard out.
+            -n count => Number of loops to run the program.  Default:  1
+            -b seconds => Polling interval in seconds.  Default:  1
+            -i [database:collection] => Name of database and collection to
+                insert the database statistics data into.
+                Default:  sysmon:mysql_perf
+            -m file => Mongo config file.  Is loaded as a python, do not
+                include the .py extension with the name.
+            -o [path]/file => Directory path and file name for output.
+            -a => Append output to output file.
+            -t email_addr email_addr2 => Enables emailing capability for an
+                option if the option allows it.  Sends output to one or more
+                email addresses.
+                -s subject_line => Subject line of email.  If none is provided
+                    then a default one will be used.
+                -u => Override the default mail command and use mailx.
+            -w => Suppress printing initial connection errors.
+            -z => Suppress standard out.
+
         -y value => A flavor id for the program lock.  To create unique lock.
-        -z => Suppress standard out.
         -v => Display version of this program.
         -h => Help and usage message.
 
         NOTE 1:  -v and/or -h overrides all other options.
+        NOTE 2:  If email is used, only a compilation of all performance
+            reports from a multiple loop run will be emailed out.
 
     Notes:
         MySQL configuration file format (config/mysql_cfg.py.TEMPLATE):
             # Configuration file:
             user = "USER"
-            passwd = "PASSWORD"
+            japd = "PSWORD"
             host = "IP_ADDRESS"
             name = "HOSTNAME"
             sid = SERVER_ID
@@ -83,12 +92,15 @@
 
             # Single Configuration file for Mongo Database Server.
             user = "USER"
-            passwd = "PASSWORD"
+            japd = "PSWORD"
             host = "IP_ADDRESS"
             name = "HOSTNAME"
             port = 27017
             conf_file = None
             auth = True
+            auth_db = "admin"
+            use_arg = True
+            use_uri = False
 
             2.)  Replica Set connection:  Same format as above, but with these
                 additional entries at the end of the configuration file:
@@ -109,14 +121,12 @@
 
 # Standard
 import sys
-import datetime
 import time
 
 # Third-party
 import json
 
 # Local
-import lib.cmds_gen as cmds_gen
 import lib.arg_parser as arg_parser
 import lib.gen_libs as gen_libs
 import lib.gen_class as gen_class
@@ -126,6 +136,9 @@ import mongo_lib.mongo_libs as mongo_libs
 import version
 
 __version__ = version.__version__
+
+# Global
+SUBJ_LINE = "MySQL_Performance"
 
 
 def help_message():
@@ -140,6 +153,35 @@ def help_message():
     """
 
     print(__doc__)
+
+
+def convert_dict(data, mail, **kwargs):
+
+    """Function:  convert_dict
+
+    Description:  Convert dictionary document to standard format and add to
+        mail body.
+
+    Arguments:
+        (input) data -> Dictionary document.
+        (input) mail -> Mail class instance.
+        (input) kwargs:
+            indent -> Level of indentation for printing.
+
+    """
+
+    data = dict(data)
+    indent = kwargs.get("indent", 0)
+    spc = " "
+
+    for key, val in data.iteritems():
+
+        if isinstance(val, dict):
+            mail.add_2_msg("{0}{1}:\n".format(spc * indent, key))
+            convert_dict(val, mail, indent=indent + 4)
+
+        else:
+            mail.add_2_msg("{0}{1}:  {2}\n".format(spc * indent, key, val))
 
 
 def mysql_stat_run(server, perf_list=None, **kwargs):
@@ -161,6 +203,7 @@ def mysql_stat_run(server, perf_list=None, **kwargs):
             ofile -> file name - Name of output file.
             no_std -> Suppress standard out.
             json_fmt -> True|False - convert output to JSON format.
+            mail -> Mail class instance.
 
     """
 
@@ -171,6 +214,7 @@ def mysql_stat_run(server, perf_list=None, **kwargs):
     ofile = kwargs.get("ofile", None)
     mode = kwargs.get("mode", "w")
     no_std = kwargs.get("no_std", False)
+    mail = kwargs.get("mail", None)
 
     if perf_list is None:
         perf_list = []
@@ -181,8 +225,7 @@ def mysql_stat_run(server, perf_list=None, **kwargs):
     server.upd_srv_stat()
     server.upd_srv_perf()
     data = {"Server": server.name,
-            "AsOf": datetime.datetime.strftime(datetime.datetime.now(),
-                                               "%Y-%m-%d %H:%M:%S"),
+            "AsOf": gen_libs.get_date() + " " + gen_libs.get_time(),
             "PerfStats": {}}
 
     for item in perf_list:
@@ -190,16 +233,14 @@ def mysql_stat_run(server, perf_list=None, **kwargs):
 
     if mongo_cfg and db_tbl:
         dbn, tbl = db_tbl.split(":")
-        mongo_libs.ins_doc(mongo_cfg, dbn, tbl, data)
+        status = mongo_libs.ins_doc(mongo_cfg, dbn, tbl, data)
+
+        if not status[0]:
+            print("Insert error:  %s" % (status[1]))
 
     if json_fmt:
         jdata = json.dumps(data, indent=indent)
-
-        if ofile:
-            gen_libs.write_file(ofile, mode, jdata)
-
-        if not no_std:
-            gen_libs.print_data(jdata)
+        _process_json(jdata, ofile, mail, mode, no_std)
 
     else:
         err_flag, err_msg = gen_libs.print_dict(data, ofile=ofile,
@@ -207,6 +248,35 @@ def mysql_stat_run(server, perf_list=None, **kwargs):
 
         if err_flag:
             print(err_msg)
+
+        if mail:
+            convert_dict(data, mail)
+
+
+def _process_json(jdata, ofile, mail, mode, no_std):
+
+    """Function:  _process_json
+
+    Description:  Process JSON formatted data.  Private function for
+        mysql_stat_run.
+
+    Arguments:
+        (input) jdata -> JSON formatted dictionary data.
+        (input) ofile -> Name of output file.
+        (input) mail -> Mail class instance.
+        (input) mode -> File write mode.
+        (input) no_std -> Suppress standard out.
+
+    """
+
+    if ofile:
+        gen_libs.write_file(ofile, mode, jdata)
+
+    if not no_std:
+        gen_libs.print_data(jdata)
+
+    if mail:
+        mail.add_2_msg(jdata)
 
 
 def mysql_stat(server, args_array, **kwargs):
@@ -224,6 +294,8 @@ def mysql_stat(server, args_array, **kwargs):
 
     """
 
+    global SUBJ_LINE
+
     args_array = dict(args_array)
     ofile = args_array.get("-o", False)
     db_tbl = args_array.get("-i", False)
@@ -231,12 +303,17 @@ def mysql_stat(server, args_array, **kwargs):
     no_std = args_array.get("-z", False)
     mode = "w"
     indent = 4
+    mail = None
 
     if args_array.get("-a", False):
         mode = "a"
 
     if args_array.get("-f", False):
         indent = None
+
+    if args_array.get("-t", None):
+        mail = gen_class.setup_mail(args_array.get("-t"),
+                                    subj=args_array.get("-s", SUBJ_LINE))
 
     # List of performance statistics to be checked.
     perf_list = ["indb_buf_data", "indb_buf_tot", "indb_buf_data_pct",
@@ -252,11 +329,17 @@ def mysql_stat(server, args_array, **kwargs):
     for item in range(0, int(args_array["-n"])):
         mysql_stat_run(server, perf_list, db_tbl=db_tbl, ofile=ofile,
                        json_fmt=json_fmt, no_std=no_std, mode=mode,
-                       indent=indent, **kwargs)
+                       indent=indent, mail=mail, **kwargs)
+
+        # Append to file after first loop.
+        mode = "a"
 
         # Do not sleep on the last loop.
         if item != int(args_array["-n"]) - 1:
             time.sleep(float(args_array["-b"]))
+
+    if mail:
+        mail.send_mail(use_mailx=args_array.get("-u", False))
 
 
 def run_program(args_array, func_dict, **kwargs):
@@ -271,21 +354,26 @@ def run_program(args_array, func_dict, **kwargs):
 
     """
 
+    mongo = None
     args_array = dict(args_array)
     func_dict = dict(func_dict)
     server = mysql_libs.create_instance(args_array["-c"], args_array["-d"],
                                         mysql_class.Server)
-    server.connect()
-    mongo = None
+    server.connect(silent=True)
 
-    if args_array.get("-m", False):
-        mongo = gen_libs.load_module(args_array["-m"], args_array["-d"])
+    if server.conn_msg and not args_array.get("-w", False):
+        print("run_program:  Error encountered on server(%s):  %s" %
+              (server.name, server.conn_msg))
 
-    # Intersect args_array and func_dict to determine which functions to call.
-    for opt in set(args_array.keys()) & set(func_dict.keys()):
-        func_dict[opt](server, args_array, class_cfg=mongo, **kwargs)
+    elif not server.conn_msg:
+        if args_array.get("-m", False):
+            mongo = gen_libs.load_module(args_array["-m"], args_array["-d"])
 
-    cmds_gen.disconnect([server])
+        # Call function(s) - intersection of command line and function dict.
+        for opt in set(args_array.keys()) & set(func_dict.keys()):
+            func_dict[opt](server, args_array, class_cfg=mongo, **kwargs)
+
+        mysql_libs.disconnect(server)
 
 
 def main():
@@ -302,6 +390,7 @@ def main():
         func_dict -> dictionary list for the function calls or other options.
         opt_def_dict -> contains options with their default values.
         opt_con_req_list -> contains the options that require other options.
+        opt_multi_list -> contains the options that will have multiple values.
         opt_req_list -> contains options that are required for the program.
         opt_val_list -> contains options which require values.
 
@@ -316,16 +405,23 @@ def main():
     file_crt_list = ["-o"]
     func_dict = {"-S": mysql_stat}
     opt_def_dict = {"-i": "sysmon:mysql_perf", "-n": "1", "-b": "1"}
-    opt_con_req_list = {"-i": ["-m", "-j"]}
+    opt_con_req_list = {"-i": ["-m", "-j"], "-s": ["-t"], "-u": ["-t"]}
+    opt_multi_list = ["-s", "-t"]
     opt_req_list = ["-c", "-d", "-b", "-n"]
-    opt_val_list = ["-c", "-d", "-b", "-i", "-m", "-n", "-o", "-y"]
+    opt_val_list = ["-c", "-d", "-b", "-i", "-m", "-n", "-o", "-s", "-t", "-y"]
 
     # Process argument list from command line.
     args_array = arg_parser.arg_parse2(cmdline.argv, opt_val_list,
-                                       opt_def_dict)
+                                       opt_def_dict, multi_val=opt_multi_list)
 
     # Add required default options and values to argument dictionary.
     args_array = arg_parser.arg_add_def(args_array, opt_def_dict, opt_req_list)
+
+    if not gen_libs.is_pos_int(int(args_array["-b"])):
+        args_array["-b"] = "1"
+
+    if not gen_libs.is_pos_int(int(args_array["-n"])):
+        args_array["-n"] = "1"
 
     if not gen_libs.help_func(args_array, __version__, help_message) \
        and not arg_parser.arg_require(args_array, opt_req_list) \