+++ test/unit/mysql_perf/main.py
@@ -68,10 +68,6 @@ class UnitTest(unittest.TestCase):
 
     Description:  Class which is a representation of a unit testing.
 
-    Super-Class:  unittest.TestCase
-
-    Sub-Classes:
-
     Methods:
         setUp -> Initialize testing environment.
         test_help_true -> Test help if returns true.
@@ -88,6 +84,12 @@ class UnitTest(unittest.TestCase):
         test_programlock_true -> Test with ProgramLock returns True.
         test_programlock_false -> Test with ProgramLock returns False.
         test_programlock_id -> Test ProgramLock with flavor ID.
+        test_interval_positive -> Test with positive interval value.
+        test_interval_zero -> Test with zero interval value.
+        test_interval_negative -> Test with negative interval value.
+        test_loop_positive -> Test with positive loop value.
+        test_loop_zero -> Test with zero loop value.
+        test_loop_negative -> Test with negative loop value.
 
     """
 
@@ -101,12 +103,26 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        self.args_array = {"-c": "CfgFile", "-d": "CfgDir"}
-        self.args_array2 = {"-c": "CfgFile", "-d": "CfgDir", "-y": "Flavor"}
+        self.args_array = {"-c": "CfgFile", "-d": "CfgDir", "-b": "1",
+                           "-n": "1"}
+        self.args_array2 = {"-c": "CfgFile", "-d": "CfgDir", "-y": "Flavor",
+                            "-b": "1", "-n": "1"}
+        self.args_array3 = {"-c": "CfgFile", "-d": "CfgDir", "-b": "1",
+                            "-n": "1"}
+        self.args_array4 = {"-c": "CfgFile", "-d": "CfgDir", "-b": "0",
+                            "-n": "1"}
+        self.args_array5 = {"-c": "CfgFile", "-d": "CfgDir", "-b": "-1",
+                            "-n": "1"}
+        self.args_array6 = {"-c": "CfgFile", "-d": "CfgDir", "-n": "1",
+                            "-b": "1"}
+        self.args_array7 = {"-c": "CfgFile", "-d": "CfgDir", "-n": "0",
+                            "-b": "1"}
+        self.args_array8 = {"-c": "CfgFile", "-d": "CfgDir", "-n": "-1",
+                            "-b": "1"}
         self.proglock = ProgramLock(["cmdline"], "FlavorID")
 
     @mock.patch("mysql_perf.gen_libs.help_func")
-    @mock.patch("mysql_perf.arg_parser.arg_parse2")
+    @mock.patch("mysql_perf.arg_parser")
     def test_help_true(self, mock_arg, mock_help):
 
         """Function:  test_help_true
@@ -117,14 +133,15 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.return_value = self.args_array
+        mock_arg.arg_parse2.return_value = self.args_array
+        mock_arg.arg_add_def.return_value = self.args_array
         mock_help.return_value = True
 
         self.assertFalse(mysql_perf.main())
 
     @mock.patch("mysql_perf.arg_parser.arg_require")
     @mock.patch("mysql_perf.gen_libs.help_func")
-    @mock.patch("mysql_perf.arg_parser.arg_parse2")
+    @mock.patch("mysql_perf.arg_parser")
     def test_help_false(self, mock_arg, mock_help, mock_req):
 
         """Function:  test_help_false
@@ -135,7 +152,8 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.return_value = self.args_array
+        mock_arg.arg_parse2.return_value = self.args_array
+        mock_arg.arg_add_def.return_value = self.args_array
         mock_help.return_value = False
         mock_req.return_value = True
 
@@ -143,7 +161,7 @@ class UnitTest(unittest.TestCase):
 
     @mock.patch("mysql_perf.arg_parser.arg_require")
     @mock.patch("mysql_perf.gen_libs.help_func")
-    @mock.patch("mysql_perf.arg_parser.arg_parse2")
+    @mock.patch("mysql_perf.arg_parser")
     def test_arg_req_true(self, mock_arg, mock_help, mock_req):
 
         """Function:  test_arg_req_true
@@ -154,7 +172,8 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.return_value = self.args_array
+        mock_arg.arg_parse2.return_value = self.args_array
+        mock_arg.arg_add_def.return_value = self.args_array
         mock_help.return_value = False
         mock_req.return_value = True
 
@@ -163,7 +182,7 @@ class UnitTest(unittest.TestCase):
     @mock.patch("mysql_perf.arg_parser.arg_cond_req")
     @mock.patch("mysql_perf.arg_parser.arg_require")
     @mock.patch("mysql_perf.gen_libs.help_func")
-    @mock.patch("mysql_perf.arg_parser.arg_parse2")
+    @mock.patch("mysql_perf.arg_parser")
     def test_arg_req_false(self, mock_arg, mock_help, mock_req, mock_cond):
 
         """Function:  test_arg_req_false
@@ -174,7 +193,8 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.return_value = self.args_array
+        mock_arg.arg_parse2.return_value = self.args_array
+        mock_arg.arg_add_def.return_value = self.args_array
         mock_help.return_value = False
         mock_req.return_value = False
         mock_cond.return_value = False
@@ -184,7 +204,7 @@ class UnitTest(unittest.TestCase):
     @mock.patch("mysql_perf.arg_parser.arg_cond_req")
     @mock.patch("mysql_perf.arg_parser.arg_require")
     @mock.patch("mysql_perf.gen_libs.help_func")
-    @mock.patch("mysql_perf.arg_parser.arg_parse2")
+    @mock.patch("mysql_perf.arg_parser")
     def test_arg_cond_false(self, mock_arg, mock_help, mock_req, mock_cond):
 
         """Function:  test_arg_cond_false
@@ -195,7 +215,8 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.return_value = self.args_array
+        mock_arg.arg_parse2.return_value = self.args_array
+        mock_arg.arg_add_def.return_value = self.args_array
         mock_help.return_value = False
         mock_req.return_value = False
         mock_cond.return_value = False
@@ -206,7 +227,7 @@ class UnitTest(unittest.TestCase):
     @mock.patch("mysql_perf.arg_parser.arg_cond_req")
     @mock.patch("mysql_perf.arg_parser.arg_require")
     @mock.patch("mysql_perf.gen_libs.help_func")
-    @mock.patch("mysql_perf.arg_parser.arg_parse2")
+    @mock.patch("mysql_perf.arg_parser")
     def test_arg_cond_true(self, mock_arg, mock_help, mock_req, mock_cond,
                            mock_dir):
 
@@ -218,7 +239,8 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.return_value = self.args_array
+        mock_arg.arg_parse2.return_value = self.args_array
+        mock_arg.arg_add_def.return_value = self.args_array
         mock_help.return_value = False
         mock_req.return_value = False
         mock_cond.return_value = True
@@ -230,7 +252,7 @@ class UnitTest(unittest.TestCase):
     @mock.patch("mysql_perf.arg_parser.arg_cond_req")
     @mock.patch("mysql_perf.arg_parser.arg_require")
     @mock.patch("mysql_perf.gen_libs.help_func")
-    @mock.patch("mysql_perf.arg_parser.arg_parse2")
+    @mock.patch("mysql_perf.arg_parser")
     def test_arg_dir_true(self, mock_arg, mock_help, mock_req, mock_cond,
                           mock_dir):
 
@@ -242,7 +264,8 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_arg.return_value = self.args_array
+        mock_arg.arg_parse2.return_value = self.args_array
+        mock_arg.arg_add_def.return_value = self.args_array
         mock_help.return_value = False
         mock_req.return_value = False
         mock_cond.return_value = True
@@ -263,6 +286,7 @@ class UnitTest(unittest.TestCase):
         """
 
         mock_arg.arg_parse2.return_value = self.args_array
+        mock_arg.arg_add_def.return_value = self.args_array
         mock_help.return_value = False
         mock_arg.arg_require.return_value = False
         mock_arg.arg_cond_req.return_value = True
@@ -284,6 +308,7 @@ class UnitTest(unittest.TestCase):
         """
 
         mock_arg.arg_parse2.return_value = self.args_array
+        mock_arg.arg_add_def.return_value = self.args_array
         mock_help.return_value = False
         mock_arg.arg_require.return_value = False
         mock_arg.arg_cond_req.return_value = True
@@ -307,6 +332,7 @@ class UnitTest(unittest.TestCase):
         """
 
         mock_arg.arg_parse2.return_value = self.args_array
+        mock_arg.arg_add_def.return_value = self.args_array
         mock_help.return_value = False
         mock_arg.arg_require.return_value = False
         mock_arg.arg_cond_req.return_value = True
@@ -331,6 +357,7 @@ class UnitTest(unittest.TestCase):
         """
 
         mock_arg.arg_parse2.return_value = self.args_array
+        mock_arg.arg_add_def.return_value = self.args_array
         mock_help.return_value = False
         mock_arg.arg_require.return_value = False
         mock_arg.arg_cond_req.return_value = True
@@ -355,6 +382,7 @@ class UnitTest(unittest.TestCase):
         """
 
         mock_arg.arg_parse2.return_value = self.args_array
+        mock_arg.arg_add_def.return_value = self.args_array
         mock_help.return_value = False
         mock_arg.arg_require.return_value = False
         mock_arg.arg_cond_req.return_value = True
@@ -379,6 +407,7 @@ class UnitTest(unittest.TestCase):
         """
 
         mock_arg.arg_parse2.return_value = self.args_array
+        mock_arg.arg_add_def.return_value = self.args_array
         mock_help.return_value = False
         mock_arg.arg_require.return_value = False
         mock_arg.arg_cond_req.return_value = True
@@ -405,6 +434,157 @@ class UnitTest(unittest.TestCase):
         """
 
         mock_arg.arg_parse2.return_value = self.args_array2
+        mock_arg.arg_add_def.return_value = self.args_array2
+        mock_help.return_value = False
+        mock_arg.arg_require.return_value = False
+        mock_arg.arg_cond_req.return_value = True
+        mock_arg.arg_dir_chk_crt.return_value = False
+        mock_arg.arg_file_chk.return_value = False
+        mock_lock.return_value = self.proglock
+
+        self.assertFalse(mysql_perf.main())
+
+    @mock.patch("mysql_perf.run_program", mock.Mock(return_value=True))
+    @mock.patch("mysql_perf.gen_class.ProgramLock")
+    @mock.patch("mysql_perf.gen_libs.help_func")
+    @mock.patch("mysql_perf.arg_parser")
+    def test_interval_positive(self, mock_arg, mock_help, mock_lock):
+
+        """Function:  test_interval_positive
+
+        Description:  Test with positive interval value.
+
+        Arguments:
+
+        """
+
+        mock_arg.arg_parse2.return_value = self.args_array3
+        mock_arg.arg_add_def.return_value = self.args_array3
+        mock_help.return_value = False
+        mock_arg.arg_require.return_value = False
+        mock_arg.arg_cond_req.return_value = True
+        mock_arg.arg_dir_chk_crt.return_value = False
+        mock_arg.arg_file_chk.return_value = False
+        mock_lock.return_value = self.proglock
+
+        self.assertFalse(mysql_perf.main())
+
+    @mock.patch("mysql_perf.run_program", mock.Mock(return_value=True))
+    @mock.patch("mysql_perf.gen_class.ProgramLock")
+    @mock.patch("mysql_perf.gen_libs.help_func")
+    @mock.patch("mysql_perf.arg_parser")
+    def test_interval_zero(self, mock_arg, mock_help, mock_lock):
+
+        """Function:  test_interval_zero
+
+        Description:  Test with zero interval value.
+
+        Arguments:
+
+        """
+
+        mock_arg.arg_parse2.return_value = self.args_array4
+        mock_arg.arg_add_def.return_value = self.args_array4
+        mock_help.return_value = False
+        mock_arg.arg_require.return_value = False
+        mock_arg.arg_cond_req.return_value = True
+        mock_arg.arg_dir_chk_crt.return_value = False
+        mock_arg.arg_file_chk.return_value = False
+        mock_lock.return_value = self.proglock
+
+        self.assertFalse(mysql_perf.main())
+
+    @mock.patch("mysql_perf.run_program", mock.Mock(return_value=True))
+    @mock.patch("mysql_perf.gen_class.ProgramLock")
+    @mock.patch("mysql_perf.gen_libs.help_func")
+    @mock.patch("mysql_perf.arg_parser")
+    def test_interval_negative(self, mock_arg, mock_help, mock_lock):
+
+        """Function:  test_interval_negative
+
+        Description:  Test with negative interval value.
+
+        Arguments:
+
+        """
+
+        mock_arg.arg_parse2.return_value = self.args_array5
+        mock_arg.arg_add_def.return_value = self.args_array5
+        mock_help.return_value = False
+        mock_arg.arg_require.return_value = False
+        mock_arg.arg_cond_req.return_value = True
+        mock_arg.arg_dir_chk_crt.return_value = False
+        mock_arg.arg_file_chk.return_value = False
+        mock_lock.return_value = self.proglock
+
+        self.assertFalse(mysql_perf.main())
+
+    @mock.patch("mysql_perf.run_program", mock.Mock(return_value=True))
+    @mock.patch("mysql_perf.gen_class.ProgramLock")
+    @mock.patch("mysql_perf.gen_libs.help_func")
+    @mock.patch("mysql_perf.arg_parser")
+    def test_loop_positive(self, mock_arg, mock_help, mock_lock):
+
+        """Function:  test_loop_positive
+
+        Description:  Test with positive loop value.
+
+        Arguments:
+
+        """
+
+        mock_arg.arg_parse2.return_value = self.args_array6
+        mock_arg.arg_add_def.return_value = self.args_array6
+        mock_help.return_value = False
+        mock_arg.arg_require.return_value = False
+        mock_arg.arg_cond_req.return_value = True
+        mock_arg.arg_dir_chk_crt.return_value = False
+        mock_arg.arg_file_chk.return_value = False
+        mock_lock.return_value = self.proglock
+
+        self.assertFalse(mysql_perf.main())
+
+    @mock.patch("mysql_perf.run_program", mock.Mock(return_value=True))
+    @mock.patch("mysql_perf.gen_class.ProgramLock")
+    @mock.patch("mysql_perf.gen_libs.help_func")
+    @mock.patch("mysql_perf.arg_parser")
+    def test_loop_zero(self, mock_arg, mock_help, mock_lock):
+
+        """Function:  test_loop_zero
+
+        Description:  Test with zero loop value.
+
+        Arguments:
+
+        """
+
+        mock_arg.arg_parse2.return_value = self.args_array7
+        mock_arg.arg_add_def.return_value = self.args_array7
+        mock_help.return_value = False
+        mock_arg.arg_require.return_value = False
+        mock_arg.arg_cond_req.return_value = True
+        mock_arg.arg_dir_chk_crt.return_value = False
+        mock_arg.arg_file_chk.return_value = False
+        mock_lock.return_value = self.proglock
+
+        self.assertFalse(mysql_perf.main())
+
+    @mock.patch("mysql_perf.run_program", mock.Mock(return_value=True))
+    @mock.patch("mysql_perf.gen_class.ProgramLock")
+    @mock.patch("mysql_perf.gen_libs.help_func")
+    @mock.patch("mysql_perf.arg_parser")
+    def test_loop_negative(self, mock_arg, mock_help, mock_lock):
+
+        """Function:  test_loop_negative
+
+        Description:  Test with negative loop value.
+
+        Arguments:
+
+        """
+
+        mock_arg.arg_parse2.return_value = self.args_array8
+        mock_arg.arg_add_def.return_value = self.args_array8
         mock_help.return_value = False
         mock_arg.arg_require.return_value = False
         mock_arg.arg_cond_req.return_value = True