+++ test/unit/mysql_perf/unit_test_run.sh
@@ -6,6 +6,8 @@
 
 echo ""
 echo "Unit testing..."
+test/unit/mysql_perf/_process_json.py
+test/unit/mysql_perf/convert_dict.py
 test/unit/mysql_perf/help_message.py
 test/unit/mysql_perf/main.py
 test/unit/mysql_perf/mysql_stat.py